+++ CHANGELOG.md
@@ -28,6 +28,9 @@ entirely to allow inference to fill in the type.
   * JSON maps are now typed as `Map<String, dynamic>` instead of
     `Map<dynamic, dynamic>`. A JSON-map is not a `HashMap` or `LinkedHashMap`
     anymore (but just a `Map`).
+* `dart:async`
+  * Add `groupBy` to `Stream`. Allows splitting a string into separate streams
+    depending on "key" property computed from the individual events.
 
 ### Dart VM